Split background module into core and background_2d
