+++ photutils/background/tests/test_background_2d.py
@@ -5,7 +5,7 @@ import numpy as np
 from numpy.testing import assert_allclose
 from astropy.tests.helper import pytest
 import itertools
-from ..core import Background2D, BackgroundIDW2D, std_blocksum
+from ..background_2d import Background2D, BackgroundIDW2D, std_blocksum
 from ...datasets import make_noise_image
 
 try:
